move route and middleware tests into subfolders
